fix mac bugs

+++ arith/babybear/src/babybearx16/babybear_neon.rs
@@ -271,7 +271,7 @@ impl From<u64> for NeonBabyBear {
     #[inline(always)]
     fn from(value: u64) -> Self {
         // BabyBear::new converts to Montgomery form
-        NeonBabyBear::pack_full(&BabyBear::new(value))
+        NeonBabyBear::pack_full(&BabyBear::new(value as u32))
     }
 }
